Added function to remove cart items, fixed cart items rendering on home page.
