+++ src/js/main.js
@@ -1,6 +1,6 @@
 import ProductData from "./ProductData.mjs";
 import ProductList from "./ProductList.mjs";
-import { updateCartCount } from "./cart";
+import { updateCartCount } from "./utils";
 
 const dataSource = new ProductData("tents");
 const element = document.querySelector(".product-list");
